fix test parameter (user correct type)

+++ src/test/java/algorithms/countinversions/SolutionTest.java
@@ -26,7 +26,7 @@ public class SolutionTest implements ArgumentsProvider {
 
   @ParameterizedTest
   @ArgumentsSource(SolutionTest.class)
-  void testInversionCountAlternative(int expected, int[] input) {
+  void testInversionCountAlternative(long expected, int[] input) {
     Assertions.assertEquals(expected, Solution.countInversionAlternative(input));
   }
 }
